+++ src/lib/app/testSvelteKitImports.ts
@@ -1,0 +1,4 @@
+export {amp, browser, dev, mode, prerendering} from '$app/env';
+export {goto, invalidate, prefetch, prefetchRoutes} from '$app/navigation';
+export {assets, base} from '$app/paths';
+export {getStores, navigating, page, session} from '$app/stores';